TL  improved the code

+++ GITProj01/src/main/java/com/nt/main/Arithmetic.java
@@ -2,11 +2,11 @@ package com.nt.main;
 
 public class Arithmetic {
 
-	public  int sum(int x,int y) {
+	public  double sum(double x,double y) {
 		return x+y;
 	}
 	
-	public    int  sub(int x,int y) {
+	public    double  sub(double x,double y) {
 		return  x-y;
 	}
 }
